Added audio and rebind menus
Implemented 2 new menus for user preferences. 1 is the audio menu that the user can fine tune their audio options, changing the volume of the music, sfx or even master audio. 2 is the rebind menu, a scene where the user can change any bound keys to their liking, or reseting to the default options. All changes made can be saved and then when re opening the program having it remember the settings