Started adding speech to text extension
Now the program asks the user on their first time playing if they want to enable tts. After that the program speaks to the user whenever they enter a scene or menu describing the function and also the user can move between buttons and labels with the directional keys. Each time the user moves to a button or label the program tells them what it does and the user can press it.